Correccion de nombres de packages. Creacion de usuarios en el inicio de la app
