+++ grails-app/controllers/arquiweb/tp1/UrlMappings.groovy
@@ -1,4 +1,4 @@
-package tp1
+package arquiweb.tp1
 
 class UrlMappings {
